+++ lib/adhearsion/configuration.rb
@@ -19,13 +19,26 @@ module Adhearsion
 
       Loquacious::Configuration.for :platform do
         root nil, :desc => "Adhearsion application root folder"
-        lib "lib", :desc => <<-__
-          Folder to include the own libraries to be used. Adhearsion loads any ruby file located into this folder during the bootstrap
-          process. Set to nil if you do not want these files to be loaded. This folder is relative to the application root folder.
-        __
-        automatically_accept_incoming_calls true, :transform => Proc.new { |v| v == 'true' }, :desc => "Adhearsion will accept automatically any inbound call"
 
-        environment :development, :transform => Proc.new { |v| v.to_sym }, :desc => "Active environment. Supported values: development, production, staging, test"
+        lib "lib", :desc => <<-__
+          Folder to include the own libraries to be used. Adhearsion loads any ruby file 
+          located into this folder during the bootstrap process. Set to nil if you do not 
+          want these files to be loaded. This folder is relative to the application root folder.
+        __
+
+        automatically_accept_incoming_calls true, :transform => Proc.new { |v| v == 'true' }, :desc => <<-__
+          Adhearsion will accept automatically any inbound call
+        __
+
+        environment :development, :transform => Proc.new { |v| v.to_sym }, :desc => <<-__
+          Active environment. Supported values: development, production, staging, test
+        __
+
+        process_name "ahn", :desc => <<-__
+          Adhearsion process name, useful to make it easier to find in the process list
+          Pro tip: set this to your application's name and you can do "killall myapp"
+          Does not work under JRuby.
+        __
 
         desc "Log configuration"
         logging {